Restore current track progress on app start
Android Media Player doesn't support seeking in most states so I added a seekTarget to PolarisMediaPlayer, which is applied when the Android Media Player is finally ready to seek.

Also allowed seeking from the PLAYBACK_COMPLETED state.

Not restoring progress if it's at the end of the track (1.f) as this leads to a silly state where pressing play resumes the track, only to instantly go back into PLAYBACK_COMPLETED. There is no way to put the player directly in the PLAYBACK_COMPLETED state so instead we ignore the progress. Pressing play just restarts the track (as it would from the PLAYBACK_COMPLETED state anyway).